basics

+++ numpy_les/02_create.py
@@ -92,7 +92,8 @@ basics.print_info(arr)
 arr = np.array([[1, 2], [3, 4]])
 arr = np.diag(arr)
 
-arr = np.empty((3, 4))
+# alloceer array zonder initialisation
+arr = np.empty((3, 4), dtype=np.int8)
 basics.print_info(arr)
 
 #
